refactor app kernel execution
