+++ marimo/_runtime/app/common.py
@@ -1,0 +1,9 @@
+# Copyright 2024 Marimo. All rights reserved.
+from typing import Any
+
+from marimo._ast.cell import CellId_t
+
+OutputsType = dict[CellId_t, Any]
+DefsType = dict[str, Any]
+
+RunOutput = tuple[OutputsType, DefsType]
